linker: resolve skippable nodes
Skippable nodes are nodes that have child nodes.

Since their children are already entitled to be symlinked individually,
there is no need to symlink those nodes, only their children.

This case is useful for directories that are wrappers of individual
configuration files, mainly the ones that are stored in the home
directory.