linker: move List method to FileSystem
List was previously in the FileInfo interface, but it didn't seem right,
since FileInfo sounds like information accumulated at once and List
wasn't behaving like that.

Also because List was moved to FileSystem and didn't belong to an
specific file info anymore, it made sense to rename it to ReadDir in
order to avoid its name to be too general.