fs: use driver abstraction for file system
It seems a better fit to have the file system be a concrete structure
that uses an abstract driver to perform operations.

This way, we can have file system methods that may depend or not on the
driver abstraction, and this abstraction may be way less bloated than if
the file system itself were to be abstract.

fs: add test package for file system

fs: move concrete file system to utility package

linker: use updated file system

cmd/plg: adapt file system calls to use a driver