cmd/plg: use external package for CLI utility
Unfortunately, https://github.com/google/subcommands is very limited and
untested, with ugly help message and lack some niceties out of the box.
I had to write a wrapper in order to have things working, it would be OK
to simply reuse the wrapper across projects if `subcommands` wasn't so
poor.

I have crafted https://github.com/gbrlsnchs/cli myself with my design of
choice, some goodies out of the box and ease in using and testing. Check
the repository for more information.

This is a total rework of the main command interface, but was totally
worth it and necessary. Now it'll be easier to introduce new features.

As a bonus, all integration tests now use the in-memory FS driver for
testing, making it even easier to prototype new functionalities.

fs: resolve slashes in filepath args of methods

This change allows someone using the `fs` package to always use forward
slashes in filepath arguments, regardless of in which OS the program
will run.

fstest: add support for absolute paths

This is not a perfect implementation, far from it. However, it is enough
for current existing tests.

config: move init method to a constructor function

It makes much more sense to initialize a new configuration through a
constructor function. By doing this, it also makes more sense now to
have functional options that set the configuration initialization,
instead of accepting maps as parameters.