cmd/plg: add option to use home as base directory
Previously, using `-basedir='$HOME'` was not a portable solution. This
commit introduces a way to correctly use the home directory with ease.

parser: resolve base directory based on mode

The parser now has two modes:

  - User mode

  - Home mode

Depending on what mode is set for a configuration, the default base
directory changes. It is always overridden by per-configuration base
directory, if it's not empty.