cmd/plg: add option to flatten target directory
It is common to have a parent directory just to group a bunch of
dotfiles from the same program. For example:

  + zsh
        +-- .zprofile
        +-- .zshrc

In this case, both files need to be symlinked in the home directory, so
we need to flatten the `zsh` directory. That's exactly what this option
does by reading all files in the parent directory.

Also, since it reads the directory, it'll override the `-targets`
option, rendering it useless. It might be soon deprecated, since listing
targets by hand is painful.