generate CLI tests using text template
This might be a code smell, but, really, it's just the target to
generate E2E CLI tests. Now it has the intelligence to use a template
when replacing texts, so I can replace specific texts with dynamic data
that can be captured somewhere in the same file.

My use case was a Windows error for directory reading: it shows the file
name, while in the Linux error it doesn't show the file name. Therefore,
the generated Windows error didn't get the proper error message,
breaking tests on CI. Now, it can be treated properly using a custom
data builder that captures the file name somehow (this is the code
smell, actually) and replace it using a text template.