cmd/plg: use multi-value option for filters
This change is due to the fact that files that have commas in their
names can't be filtered using comma-separated options.

Before:

  plg init -include foo,bar

After:

  plg init -include foo -include bar

Yes, it's more verbose, but it's safer. What if a user accidentally
creates a file with a comma in its name and then adds it to Pilgo? They
won't be able to remove it via commands because the file won't be
filterable, and they'll have to manually edit the configuration file.